separating groupChar to new submodule
