+++ pkg/arg_list_comp/charlist.go
@@ -15,7 +15,7 @@ type charList struct {
 }
 
 func newAlcCharList(s string) (cl *charList, err error) {
-	prefix, comprehension, suffix, err := parts(s, "[")
+	prefix, comprehension, suffix, err := groupChar("[").Parts(s)
 	if err != nil {
 		return nil, err
 	}
